% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.vdj.R
\name{add.vdj}
\alias{add.vdj}
\title{Add CITE-seq antibody-derived tags (ADT)}
\usage{
add.vdj(x = NULL, vdj.data = "data.frame")
}
\arguments{
\item{x}{An object of class scSeqR.}

\item{adt.data}{A data frame containing ADT counts for cells.}
}
\value{
An object of class scSeqR
}
\description{
This function takes a data frame of ADT values per cell and adds it to the scSeqR object.
}
\examples{
\dontrun{
my.obj <- add.adt(my.obj, adt.data = adt.data)
}

}
