% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt.rna.merge.R
\name{adt.rna.merge}
\alias{adt.rna.merge}
\title{Create an object of class scSeqR.}
\usage{
adt.rna.merge(x = NULL, adt.data = "raw")
}
\arguments{
\item{x}{A data frame containing gene counts for cells.}
}
\value{
An object of class scSeqR
}
\description{
This function takes data frame and makes an object of class scSeqR.
}
\examples{
\dontrun{
make.obj(my.data)
}

}
