% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.clust.R
\name{change.clust}
\alias{change.clust}
\title{Assign the optimal number of clusters to hierarchical clustering.}
\usage{
change.clust(x = NULL, change.clust = 0, to.clust = 0)
}
\arguments{
\item{x}{An object of class scSeqR.}

\item{clust.num}{Number of clusters}

\item{clust.type}{Choose from "tsne","pca" or "distance", defult = "tsne".}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and finds optimal number of clusters based on three methods.
}
\examples{
\dontrun{
assign.clust(my.obj, clust.num = 7)
}
}
