% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust.avg.exp.R
\name{clust.avg.exp}
\alias{clust.avg.exp}
\title{Scale data}
\usage{
clust.avg.exp(x = NULL)
}
\arguments{
\item{x}{An object of class scSeqR.}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and scales the normalized data.
}
\examples{
\dontrun{
scale.data(my.obj)
}
}
