% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust.rm.R
\name{clust.rm}
\alias{clust.rm}
\title{Create an object of class scSeqR.}
\usage{
clust.rm(x = NULL, clust.to.rm = "numeric")
}
\arguments{
\item{x}{A data frame containing gene counts for cells.}
}
\value{
An object of class scSeqR
}
\description{
This function takes data frame and makes an object of class scSeqR.
}
\examples{
\dontrun{
make.obj(my.data)
}

}
