% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.plot.R
\name{cluster.plot}
\alias{cluster.plot}
\title{Plot nGenes, UMIs and perecent mito}
\usage{
cluster.plot(x = NULL, cell.size = 1, plot.type = "tsne",
  cell.color = "black", clustered = TRUE)
}
\arguments{
\item{x}{An object of class scSeqR.}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and creats plot.
}
\examples{
\dontrun{
tsne.plot(my.obj)
}
}
