% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.exp.R
\name{diff.exp}
\alias{diff.exp}
\title{Create an object of class scSeqR.}
\usage{
diff.exp(x = NULL, de.by = "clusters", cond.1 = "array",
  cond.2 = "array")
}
\arguments{
\item{x}{A data frame containing gene counts for cells.}
}
\value{
An object of class scSeqR
}
\description{
This function takes data frame and makes an object of class scSeqR.
}
\examples{
\dontrun{
make.obj(my.data)
}

}
