% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down.sample.R
\name{down.sample}
\alias{down.sample}
\title{Down sample conditions}
\usage{
down.sample(x = NULL)
}
\arguments{
\item{x}{An object of class scSeqR.}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and down samples the condition to have equal number of cells in each condition.
}
\examples{
\dontrun{
my.obj <- down.sample(my.obj)
}
}
