% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.cells.R
\name{filter.cells}
\alias{filter.cells}
\title{Filter cells}
\usage{
\method{filter}{cells}(x = NULL, min.mito = 0, max.mito = 1,
  min.genes = 0, max.genes = Inf, min.umis = 0, max.umis = Inf)
}
\arguments{
\item{x}{An object of class scSeqR.}

\item{min.mito}{Min ratio of mit genes, defult = 0.}

\item{max.mito}{Max ratio of mit genes, defult = 1.}

\item{min.genes}{Min number genes per cell, defult = 0.}

\item{max.genes}{Max number genes per cell, defult = Inf.}

\item{min.umis}{Min number UMIs per cell, defult = 0.}

\item{max.umis}{Max number UMIs per cell, defult = Inf.}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and filters the raw data based on the number of UMIs, genes per cell and percentage of mitochondrial genes per cell.
}
\examples{
\dontrun{
filter.cells(my.obj,
   min.mito = 0,
   max.mito = 1,
   min.genes = 0,
   max.genes = Inf,
   min.umis = 0,
   max.umis = Inf)
}
}
