% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.marker.R
\name{find.marker}
\alias{find.marker}
\title{Create an object of class scSeqR.}
\usage{
find.marker(x = NULL, fold.change = 1.5)
}
\arguments{
\item{x}{A data frame containing gene counts for cells.}
}
\value{
An object of class scSeqR
}
\description{
This function takes data frame and makes an object of class scSeqR.
}
\examples{
\dontrun{
make.obj(my.data)
}

}
