% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.markers.R
\name{find.markers}
\alias{find.markers}
\title{Create an object of class scSeqR.}
\usage{
find.markers(x = NULL, data.type = "tsne", fold.change = 2,
  padjval = 0.1, Inf.FCs = FALSE, positive = TRUE)
}
\arguments{
\item{x}{A data frame containing gene counts for cells.}
}
\value{
An object of class scSeqR
}
\description{
This function takes data frame and makes an object of class scSeqR.
}
\examples{
\dontrun{
make.obj(my.data)
}

}
