% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.opt.pcs.R
\name{find.opt.pcs}
\alias{find.opt.pcs}
\title{Find optimal number of PCs for clustering}
\usage{
find.opt.pcs(x = NULL, pcs.in.plot = 50)
}
\arguments{
\item{x}{An object of class scSeqR.}

\item{pcs.in.plot}{Number of PCs to show in plot, defult = 50.}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and finds optimal number of PCs for clustering.
}
\examples{
\dontrun{
find.opt.pcs(my.obj)
}

}
