% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.stats.plot.R
\name{gene.stats.plot}
\alias{gene.stats.plot}
\title{Scale data}
\usage{
gene.stats.plot(x = NULL)
}
\arguments{
\item{x}{An object of class scSeqR.}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and scales the normalized data.
}
\examples{
\dontrun{
gene.stats.plot(my.obj)
}
}
