% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.plor.R
\name{heatmap.plot}
\alias{heatmap.plot}
\title{Create an object of class scSeqR.}
\usage{
heatmap.plot(x = NULL, gene = "NULL", plot.data.type = "tsne",
  clust.dim = 2, cluster.by = "clusters", heat.colors = c("blue", "white",
  "red"))
}
\arguments{
\item{x}{A data frame containing gene counts for cells.}
}
\value{
An object of class scSeqR
}
\description{
This function takes data frame and makes an object of class scSeqR.
}
\examples{
\dontrun{
make.obj(my.data)
}
}
