% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imm.gen.R
\name{imm.gen}
\alias{imm.gen}
\title{Create heatmaps for genes in clusters or conditions.}
\usage{
imm.gen(immgen.data = "rna", gene = "NULL", plot.type = "heatmap",
  heat.colors = c("blue", "white", "red"))
}
\arguments{
\item{gene}{A set of gene names to be heatmapped.}

\item{heat.colors}{Colors for heatmap, defult = c("blue" ,"white", "red").}

\item{x}{A data frame containing gene counts for cells.}

\item{cluster.by}{Choose from "clusters" or "conditions", defult = "clusters".}

\item{cluster.rows}{If set to FALSE the genes would not be clustered, defult = TRUE.}

\item{scale}{Choose from "row" or "column", defult = "row".}
}
\value{
An object of class scSeqR
}
\description{
This function takes an object of class scSeqR and genes and provides a heatmap.
}
\examples{
\dontrun{
MyGenes <- c("SOD1","CD7")
MyGenes <- top.markers(marker.genes, topde = 10, min.base.mean = 0.8)
heatmap.plot (my.obj, gene = MyGenes, cluster.by = "clusters", cluster.rows = T)
}
}
