% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load10x.R
\name{load10x}
\alias{load10x}
\title{Load 10X data as dgTMatrix}
\usage{
load10x(dir.10x = NULL)
}
\arguments{
\item{dir.10x}{A directory that includes the 10X barcodes.tsv, genes.tsv and matrix.mtx files.}
}
\value{
The dgTMatrix matrix.
}
\description{
This function takes 10X data files barcodes.tsv, genes.tsv and matrix.mtx and converts them to proper matrix file for scSeqR.
}
\examples{
\dontrun{
load10x('/hg19')
}
}
