% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.gene.model.R
\name{make.gene.model}
\alias{make.gene.model}
\title{Scale data}
\usage{
make.gene.model(x = NULL, dispersion.limit = 1.5, base.mean.rank = 500,
  non.sig.col = "darkgray", right.sig.col = "chartreuse3",
  left.sig.col = "cadetblue3", disp.line.col = "black",
  rank.line.col = "red", cell.size = 1.75, cell.transparency = 0.5,
  no.mito.model = T)
}
\arguments{
\item{x}{An object of class scSeqR.}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and scales the normalized data.
}
\examples{
\dontrun{
gene.stats.plot(my.obj)
}
}
