% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.obj.R
\name{make.obj}
\alias{make.obj}
\title{Create an object of class scSeqR.}
\usage{
make.obj(x = NULL)
}
\arguments{
\item{x}{A data frame containing gene counts for cells.}
}
\value{
An object of class scSeqR
}
\description{
This function takes data frame and makes an object of class scSeqR.
}
\examples{
\dontrun{
make.obj(my.data)
}
}
