% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man.assign.clust.R
\name{man.assign.clust}
\alias{man.assign.clust}
\title{Assign the optimal number of clusters to hierarchical clustering.}
\usage{
man.assign.clust(x = NULL, clust.num = 0)
}
\arguments{
\item{x}{An object of class scSeqR.}

\item{clust.num}{Number of clusters}

\item{clust.type}{Choose from "tsne","pca" or "distance", defult = "tsne".}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and finds optimal number of clusters based on three methods.
}
\examples{
\dontrun{
assign.clust(my.obj, clust.num = 7)
}
}
