% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.adt.R
\name{merge.adt}
\alias{merge.adt}
\title{Create an object of class scSeqR.}
\usage{
\method{merge}{adt}(x = NULL, adt.data = "raw")
}
\arguments{
\item{x}{A data frame containing gene counts for cells.}
}
\value{
An object of class scSeqR
}
\description{
This function takes data frame and makes an object of class scSeqR.
}
\examples{
\dontrun{
make.obj(my.data)
}

}
