% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.adt.R
\name{norm.adt}
\alias{norm.adt}
\title{Create an object of class scSeqR.}
\usage{
norm.adt(x = NULL)
}
\arguments{
\item{x}{A data frame containing gene counts for cells.}
}
\value{
An object of class scSeqR
}
\description{
This function takes data frame and makes an object of class scSeqR.
}
\examples{
\dontrun{
make.obj(my.data)
}

}
