% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm}
\alias{norm}
\title{Normalize data}
\usage{
norm(x = NULL, norm.method = "ranked.glsf", top.rank = 500)
}
\arguments{
\item{x}{An object of class scSeqR.}

\item{norm.method}{Choose a normalization method, there are three option currently.
Choose from "global.glsf", "ranked.glsf" or "spike.in", defult = "ranked.glsf".}

\item{top.rank}{If the method is set to "ranked.glsf", you need to set top number of genes sorted based on global base mean, defult = 500.}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and normalized the data based on "global.glsf", "ranked.glsf" or "spike.in" methods.
}
\examples{
\dontrun{
norm(my.obj, "ranked.glsf", top.rank = 500)
}
}
