% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.clust.num.R
\name{opt.clust.num}
\alias{opt.clust.num}
\title{Find optimal number of clusters.}
\usage{
opt.clust.num(x = NULL, max.clust = 15, gap.stat.nboot = 100,
  verbose = TRUE)
}
\arguments{
\item{x}{An object of class scSeqR.}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and finds optimal number of clusters based on three methods.
}
\examples{
\dontrun{
pot.clust.num(my.obj)
}
}
