% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassign.clust.R
\name{reassign.clust}
\alias{reassign.clust}
\title{Assign the optimal number of clusters to hierarchical clustering.}
\usage{
reassign.clust(x = NULL, change.clust = 0, to.clust = 0,
  clust.type = "tsne")
}
\arguments{
\item{x}{An object of class scSeqR.}

\item{clust.type}{Choose from "tsne","pca" or "distance", defult = "tsne".}

\item{clust.num}{Number of clusters}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and finds optimal number of clusters based on three methods.
}
\examples{
\dontrun{
assign.clust(my.obj, clust.num = 7)
}
}
