% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.pca.R
\name{run.pca}
\alias{run.pca}
\title{Cluster based on tSNE}
\usage{
run.pca(x = NULL, clust.method = "base.mean.rank", top.rank = 500,
  gene.list = "my_model_genes.txt")
}
\arguments{
\item{x}{An object of class scSeqR.}

\item{clust.method}{A method to choose your genes for clustering. Choose from "base.mean.rank" or "gene.model", defult is "base.mean.rank".}

\item{top.rank}{Number of ranked genes you want to use for clustering, defult = 500.}

\item{clust.type}{Choose from "tsne", "pca" or "distance".}

\item{dist.method}{Choose "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski", defult = "euclidean".}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and clusters the normalized data.
}
\examples{
\dontrun{
cluster.data(my.obj, clust.method = "base.mean.rank", top.rank = 500, clust.dim = 3)
}
}
