% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.umap.R
\name{run.umap}
\alias{run.umap}
\title{Run UMAP on PCA data}
\usage{
run.umap(x = NULL, dims = 1:20)
}
\arguments{
\item{x}{An object of class scSeqR.}

\item{dims}{PC dimentions to be used for UMAP analysis.}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and runs UMAP on PCA data.
}
\examples{
\dontrun{
my.obj <- run.umap(my.obj, dims = 1:10)
}
}
