% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.data.R
\name{scale.data}
\alias{scale.data}
\title{Scale data}
\usage{
\method{scale}{data}(x = NULL)
}
\arguments{
\item{x}{An object of class scSeqR.}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and scales the normalized data.
}
\examples{
\dontrun{
scale.data(my.obj)
}
}
