% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.genes.R
\name{stats.genes}
\alias{stats.genes}
\title{Plot nGenes, UMIs and perecent mito}
\usage{
stats.genes(x = NULL)
}
\arguments{
\item{x}{An object of class scSeqR.}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and creats plot.
}
\examples{
\dontrun{
plot.stats(my.obj)
}
}
