% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.plot.R
\name{stats.plot}
\alias{stats.plot}
\title{Plot nGenes, UMIs and perecent mito}
\usage{
stats.plot(x = NULL, cell.color = "slategray3", box.color = "red",
  cell.size = 1)
}
\arguments{
\item{x}{An object of class scSeqR.}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and creats plot.
}
\examples{
\dontrun{
plot.stats(my.obj)
}
}
