% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top.markers.R
\name{top.markers}
\alias{top.markers}
\title{Load 10X data as data.frame}
\usage{
top.markers(x = NULL, topde = 5, min.base.mean = 0.5)
}
\arguments{
\item{dir.10x}{A directory that includes the 10X barcodes.tsv, genes.tsv and matrix.mtx files.}

\item{gene.name}{Should be either geneSymbol or ensembleID.}
}
\value{
The data frame object
}
\description{
This function takes 10X data files barcodes.tsv, genes.tsv and matrix.mtx and converts them to proper matrix file for scSeqR.
}
\examples{
\dontrun{
load10x("/hg19", gene.name = "geneSymbol")
}
}
