% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.plot.R
\name{tsne.plot}
\alias{tsne.plot}
\title{Plot nGenes, UMIs and perecent mito}
\usage{
tsne.plot(x = NULL)
}
\arguments{
\item{x}{An object of class scSeqR.}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and creats plot.
}
\examples{
\dontrun{
tsne.plot(my.obj)
}
}
