% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMIs.genes.mito.R
\name{UMIs.genes.mito}
\alias{UMIs.genes.mito}
\title{Calculate the number of UMIs and genes per cell and percentage of mitochondrial genes per cell.}
\usage{
UMIs.genes.mito(x = NULL)
}
\arguments{
\item{x}{A data frame containing gene counts for cells.}
}
\value{
The data frame object
}
\description{
This function takes data frame and calculates the number of UMIs, genes per cell and percentage of mitochondrial genes per cell.
}
\examples{
\dontrun{
UMIs.genes.mit(my.data)
}
}
