% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.ma.plot.R
\name{volcano.ma.plot}
\alias{volcano.ma.plot}
\title{Create an object of class scSeqR.}
\usage{
volcano.ma.plot(x = NULL, sig.value = "padj", sig.line = 0.1,
  plot.type = "volcano", x.limit = 2, y.limit = 2, limit.force = F,
  scale.ax = T, dot.size = 1.75, dot.transparency = 0.5,
  dot.col = c("#E64B35", "#3182bd", "#636363"), interactive = TRUE,
  out.name = "plot")
}
\arguments{
\item{x}{A data frame containing gene counts for cells.}
}
\value{
An object of class scSeqR
}
\description{
This function takes data frame and makes an object of class scSeqR.
}
\examples{
\dontrun{
make.obj(my.data)
}

}
