% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.clust.R
\name{assign.clust}
\alias{assign.clust}
\title{Find optimal number of clusters.}
\usage{
assign.clust(x = NULL, clust.num = 0)
}
\arguments{
\item{x}{An object of class scSeqR.}

\item{clust.num}{Number of clusters}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and finds optimal number of clusters based on three methods.
}
\examples{
\dontrun{
assign.clust(my.obj, clust.num = 7)
}
}
