% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.data.R
\name{cluster.data}
\alias{cluster.data}
\title{Cluster based on tSNE}
\usage{
cluster.data(x = NULL, clust.method = "base.mean.rank", top.rank = 500,
  clust.dim = 2)
}
\arguments{
\item{x}{An object of class scSeqR.}

\item{clust.method}{a method to choose your genes for clustering. There are three options "base.mean.rank","dispersed.genes","both" defults is "base.mean.rank".}

\item{top.rank}{number of ranked genes you want to use for clustering, defult = 500.}

\item{clust.dim}{number of dimentions you want to consider for clustering, defult =2.}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and clusters the normalized data.
}
\examples{
\dontrun{
cluster.data(my.obj, clust.method = "base.mean.rank", top.rank = 500, clust.dim = 3)
}
}
