% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.data.R
\name{cluster.data}
\alias{cluster.data}
\title{Cluster based on tSNE}
\usage{
cluster.data(x = NULL, clust.method = "base.mean.rank", top.rank = 500,
  gene.list = NULL, clust.dim = 2, clust.type = "tsne",
  dist.method = "euclidean")
}
\arguments{
\item{x}{An object of class scSeqR.}

\item{clust.method}{A method to choose your genes for clustering. There are three options "base.mean.rank","dispersed.genes","both" or "my.genes",defults is "base.mean.rank".}

\item{top.rank}{Number of ranked genes you want to use for clustering, defult = 500.}

\item{clust.dim}{Number of dimentions you want to consider for clustering, defult =2.}

\item{clust.type}{Choose from "tsne", "pca" or "distance".}

\item{dist.method}{Choose "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski", defult = "euclidean".}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and clusters the normalized data.
}
\examples{
\dontrun{
cluster.data(my.obj, clust.method = "base.mean.rank", top.rank = 500, clust.dim = 3)
}
}
