% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.stats.R
\name{gene.stats}
\alias{gene.stats}
\title{Scale data}
\usage{
gene.stats(x = NULL, which.data = "raw.data")
}
\arguments{
\item{x}{An object of class scSeqR.}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and scales the normalized data.
}
\examples{
\dontrun{
gene.stats(my.obj)
}
}
