% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.plot.R
\name{heatmap.plot}
\alias{heatmap.plot}
\title{Create an object of class scSeqR.}
\usage{
heatmap.plot(x = NULL, gene = "NULL", cluster.by = "clusters",
  cluster.rows = T, scale = "row", heat.colors = c("blue", "white",
  "red"))
}
\arguments{
\item{x}{A data frame containing gene counts for cells.}
}
\value{
An object of class scSeqR
}
\description{
This function takes data frame and makes an object of class scSeqR.
}
\examples{
\dontrun{
make.obj(my.data)
}
}
