% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.pc.tsne.R
\name{run.pc.tsne}
\alias{run.pc.tsne}
\title{Run tSNE on PCA data}
\usage{
run.pc.tsne(x = NULL, dims = 1:20)
}
\arguments{
\item{x}{An object of class scSeqR.}

\item{dims}{PC dimentions to be used for tSNE analysis.}
}
\value{
An object of class scSeqR.
}
\description{
This function takes an object of class scSeqR and runs tSNE on PCA data.
}
\examples{
\dontrun{
my.obj <- run.pc.tsne(my.obj, dims = 1:10)
}
}
